+++ HISTORY.md
@@ -1,5 +1,6 @@
 Release 1.0.0-alpha2
 
+- begin detecting and reporting on changed client capabilities
 - add --yes arg to support automatic yes to --clean y/n prompt
 - refactors; bump tox version in requirements
 - remove menu mode code that is no longer used in WLAN Pi v2 builds